Logger Addition and Changes
Work to introduce a logger to log calls made to the API and errors generated through the SQL elements.

Changes have also been made to the Assistant Filters for the swagger page. Required tag has been removed and replaced with an operation filter that sets the filter as required in the swagger page.

Added some additional code to set the Swagger endpoint depending on the environment running.

Added the bool data type to the Model Validation Service.