Login Attempt Logs
Added a table to the database to store login attempts. Added a new method in Audit History Service to log the attempt.

Made some logging changes to handle  SQL logs correctly as Summaries were missing.

Changed some code to allow for nulls to be correctly passed to the SQL query.