Token Endpoint
Under Kyle's expert eye I have written the first set of changes to the API to make it a proper web application.

The code has been tested and works as expected with swagger page generation and set up.

Need to alter how passwords are handled. Notes have been taken and this will be addressed in another commit.