Token Alterations
Made changes to the TokenService to remove functions and converters into their own classes. This is to ensure the service is only for SQL access methods.

Updated Swagger.CSS to remove the "Try It Out" button.

Added a function to hash string values. Passwords are now hashed once extracted from header and checked against hashed password in the database.