Connection Port
Added the Port column to the ERD.

Added the Port column to the database setup SQL.

Updated the models to have a "Port" property.

Updated the SQL to reference this column.

Updated the service to reference the new property.

Updated the controller to reference the new property.

Tested all code and works with no issue.